requirements.txt
flask
flask-login

requests
beautifulsoup4

opencv-contrib-python-headless
pillow

keras
tensorflow